Rescope MasterProxy.Application into Master proxy and turn it into a Supervisor rather than an OTP app. Take in and store in application environment runtime config offered by the caller, overriding compile-time configs
