+++ mix.exs
@@ -23,8 +23,7 @@ defmodule MasterProxy.MixProject do
   # Run "mix help compile.app" to learn about applications.
   def application do
     [
-      extra_applications: [:logger],
-      mod: {MasterProxy.Application, []}
+      extra_applications: [:logger]
     ]
   end
